refactor: refactor file structure and clean up import, comments and links
