+++ src/App.jsx
@@ -1,29 +1,20 @@
 import { BrowserRouter } from "react-router-dom";
 import { useEffect, useState } from "react";
-import { About, Contact, Hero, Navbar, Projects, MoonCanvas, StarsCanvas, Footer } from './components';
+import { About, Contact, Hero, Navbar, Projects, StarsCanvas, Footer } from './components';
 import { MotionConfig } from "framer-motion";
 
 const App = () => {
   const [isMobile, setIsMobile] = useState(false);
 
   useEffect(() => {
-    // Add a listener for changes to the screen size
     const mediaQuery = window.matchMedia("(max-width: 500px)");
-
-    // Set the initial value of the `isMobile` state variable
     setIsMobile(mediaQuery.matches);
-
-    // Define a callback function to handle changes to the media query
     const handleMediaQueryChange = (event) => {
  
       setIsMobile(event.matches);
       
     };
-
-    // Add the callback function as a listener for changes to the media query
     mediaQuery.addEventListener("change", handleMediaQueryChange);
-
-    // Remove the listener when the component is unmounted
     return () => {
       mediaQuery.removeEventListener("change", handleMediaQueryChange);
     };
@@ -38,17 +29,11 @@ const App = () => {
           <Navbar />
           <StarsCanvas />
 
-          <Hero  isMobile={isMobile}/>
-          {/* <div className="relative snap-y snap-mandatory"> */}
+          <Hero />
           <About isMobile={isMobile} />
-          {/* <MoonCanvas /> */}
-          <Projects   isMobile={isMobile}/>
-          <Contact  isMobile={isMobile} />
-          {/* </div> */}
-
+          <Projects  />
+          <Contact />
           <Footer />
-
-          {/* <div className="relative z-0"></div> */}
         </div>
       </MotionConfig>
     </BrowserRouter>
