+++ src/components/Footer.jsx
@@ -1,12 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-
-
-import { navLinks } from "../constants";
-import { linkedin, github, email } from "../assets";
-
 import SocialLinks from "./SocialLinks";
-
 
 const Footer = () => {
     const currentYear = new Date().getFullYear();
