+++ src/components/Hero.jsx
@@ -1,26 +1,13 @@
 import {motion } from 'framer-motion';
-import { slideIn } from "../utils/motion";
-
-import { Canvas } from '@react-three/fiber';
-import { useScroll, animated } from "@react-spring/web";
-
 import { MoonCanvas } from './canvas';
 
 
-const Hero = ({isMobile}) => {
-  const { scrollYProgress } = useScroll(); 
- 
+const Hero = () => {
 
   return (
-    <>
-      {/* <section className="relative w-full h-screen mx-auto">*/}
       <section
         className={`relative sm:h-screen pt-20 sm:pt-0 mx-auto flex sm:flex-row flex-col-reverse justify-center items-center snap-center snap-always`}
       >
-        {/* <motion.div
-          variants={slideIn("left", "tween", 0.2, 1)}
-          className="flex-[0.75] flex gap-10 xl:flex-row flex-row flex-nowrap items-center justify-center"
-        > */}
         <span className="hash-span" id="hero"></span>
         <div className="flex-initial sm:max-w-[50vw] sm:min-w-[50vw] max-w-[90vw] flex flex-row flex-nowrap sm:place-items-center sm:place-content-center -ml-5 sm:ml-0 items-start justify-start">
           <div
@@ -54,10 +41,9 @@ const Hero = ({isMobile}) => {
             </a>
           </div>
         </div>
-        {/* </motion.div> */}
 
         <div className="flex-1 w-full sm:max-w-[50vw] min-w-[50wh] 2xl:h-min-[820px] 2xl:h-[820px] lg:h-[650px] md:h-[550px] h-[350px]">
-          <MoonCanvas isMobile={isMobile} />
+          <MoonCanvas/>
         </div>
 
         <div className="invisible sm:visible absolute xs:bottom-10 bottom-32 w-full flex justify-center items-center drop-shadow-md z-20">
@@ -78,10 +64,7 @@ const Hero = ({isMobile}) => {
           </a>
         </div>
       </section>
-
-      {/* </section> */}
-    </>
   );
 }
 
-export default Hero
+export default Hero;
