+++ src/components/SocialLinks.jsx
@@ -1,12 +1,11 @@
-import React, { useEffect, useState } from "react";
 import { linkedin, github, email } from "../assets";
 
 const SocialLinks = () => (
   <div className="flex flex-row gap-5 items-center justify-center">
-    <div
-      onClick={() =>
-        window.open("https://linkedin.com/in/eva-goetzke", "_blank")
-      }
+    <a
+      href="https://linkedin.com/in/eva-goetzke"
+      target="_blank"
+      rel="noreferrer"
       className="w-6 h-6 flex justify-center items-center cursor-pointer drop-shadow-md"
     >
       <img
@@ -14,14 +13,15 @@ const SocialLinks = () => (
         alt="linkedin"
         className="w-full h-full object-contain"
       />
-    </div>
-    <div
-      onClick={() => window.open("https://github.com/esgoet", "_blank")}
+    </a>
+    <a
+      href="https://github.com/esgoet"
+      target="_blank"
+      rel="noreferrer"
       className="w-7 h-7 flex justify-center items-center cursor-pointer drop-shadow-md"
     >
       <img src={github} alt="github" className="w-full h-full object-contain" />
-    </div>
-
+    </a>
     <a
       href="#contact"
       className="w-7 h-7 flex justify-center items-center cursor-pointer drop-shadow-md"
