+++ src/components/index.js
@@ -2,7 +2,7 @@ import { SatelliteCanvas, StarsCanvas, MoonCanvas } from './canvas';
 import Hero from './Hero';
 import Navbar from './Navbar';
 import About from './About';
-import Projects from './Projects';
+import Projects from './project/Projects.jsx';
 import Contact from './Contact';
 import Footer from './Footer';
 import SocialLinks from './SocialLinks';
