+++ src/components/project/GalleryScrollIndicator.jsx
