+++ src/hoc/SectionWrapper.jsx
@@ -1,4 +1,4 @@
-import {useState, useEffect, useRef, useContext} from "react";
+import {useState, useEffect} from "react";
 import { motion } from "framer-motion";
